Rename duplicate folder
